+++ components/keyscreen/impl/src/main/java/com/flipperdevices/keyscreen/impl/composable/actions/ComposableShare.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import com.flipperdevices.keyscreen.impl.R
 import com.flipperdevices.keyscreen.impl.composable.actions.common.ComposableActionRow
 import com.flipperdevices.keyscreen.impl.composable.actions.common.ComposableActionRowInProgress
-import com.flipperdevices.keyscreen.impl.model.ShareState
+import com.flipperdevices.keyscreen.model.ShareState
 import com.flipperdevices.core.ui.res.R as DesignSystem
 
 @Composable