+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 # 1.5.2 - In Progress
 
 - [FIX] Add confirm delete dialog and some small fixes
-- [FIX] Migrate key emulate to new module
 - [Feature] FapHub MVPv0
 - [Feature] Update README with Fdroid badge
 - [Refactor] Migrate all feature components modules from KSP to Anvil
 - [Refactor] Migrate to Ktorfit
+- [Refactor] Migrate key emulate to new module
+- [Refactor] Key Screen state in API and KeyCard with state
 
 # 1.5.1